dansah/requirements.txt
Django==4.2.1
django_storages==1.13.2
djangorestframework==3.14.0
